优化代码结构，base model相关统一放在models.py下
