+++ python/core/manifest/manifest_decorator.py
@@ -12,72 +12,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 import inspect
-import re
-from enum import Enum
-from typing import Any, Dict, List, Optional
+from typing import Any, Dict, Optional
 
 from pydantic.v1 import BaseModel as PydanticBaseModel # noqa: F403 # type: ignore
 from pydantic.v1 import create_model
 
-from dataclasses_json import config
-from dataclasses import dataclass
-from dataclasses import field as Field
-from dataclasses_json import DataClassJsonMixin
-
 from appbuilder.core.manifest.manifest_function_signature import get_signature
-
-class BaseModel(DataClassJsonMixin):
-    """Base model for all models."""
-
-    pass
-
-FUNCTION_NAME_REGEX = r"^[0-9A-Za-z_\u4e00-\u9fff]*$"
-FUNCTION_PARAM_NAME_REGEX = r"^[0-9A-Za-z_\u4e00-\u9fff]*$"
-FUNCTION_FLAG = "__pf_function__"
-
-
-def validate_function_name(value: Optional[str]) -> None:
-    """
-    Validates that the function name is valid.
-
-    Valid function names are non-empty and
-    match the regex: [0-9A-Za-z_]*
-
-    :param value: The function name to validate.
-
-    :raises ValueError: If the function name is invalid.
-    """
-    if not value:
-        raise ValueError("The function name cannot be `None` or empty")
-
-    if not re.match(FUNCTION_NAME_REGEX, value):
-        raise ValueError(
-            f"Invalid function name: {value}. Function "
-            f"names may only contain ASCII letters, "
-            f"digits, and underscores."
-        )
-
-
-def validate_function_param_name(value: Optional[str]) -> None:
-    """
-    Validates that the function parameter name is valid.
-
-    Valid function parameter names are non-empty and
-    match the regex: [0-9A-Za-z_]*
-
-    :param value: The function parameter name to validate.
-
-    :raises ValueError: If the function parameter name is invalid.
-    """
-    if not value:
-        raise ValueError("The function parameter name cannot be `None` or empty")
-
-    if not re.match(FUNCTION_PARAM_NAME_REGEX, value):
-        raise ValueError(
-            f"Invalid function parameter name: {value}. Function parameter "
-            f"names may only contain ASCII letters, digits, and underscores."
-        )
-
+from appbuilder.core.manifest.models import ParameterViewKind, ParameterView, ManifestView
 
 # The following two functions are here to allow dynamically updating function description.
 # Check tool/builtin/openapi_plugin.py for example.
@@ -126,148 +67,6 @@ def get_function_schema_with_inspect(method):
         ).schema()["properties"]
     else:  # method has no arguments
         return None
-
-def _exclude(value):
-    return True
-
-class ParameterViewKind(str, Enum):
-    """
-    The kind of a parameter.
-
-
-    Args:
-        ARGUMENT: The parameter is a function argument.
-        RETURN: The parameter is a function return.
-        PATH: The parameter is a OpenAPI path parameter.
-        QUERY: The parameter is a OpenAPI query parameter.
-        HEADER: The parameter is a OpenAPI header parameter.
-        COOKIE: The parameter is a OpenAPI cookie parameter.
-    """
-
-    ARGUMENT = "argument"
-    RETURN = "return"
-    PATH = "path"
-    QUERY = "query"
-    HEADER = "header"
-    COOKIE = "cookie"
-    BODY = "body"
-
-@dataclass
-class ParameterView(BaseModel):
-    """
-    View of a function parameter.
-
-    Args:
-        name (str): The name of the parameter.
-        description (str): The description of the parameter.
-        default_value (str): The default value of the parameter.
-        type_ (str): The type of the parameter.
-        type_object (Any): The type object of the parameter.
-        type_schema (Dict): The json schema of the parameter.
-        required (bool): Whether the parameter is required.
-        example (str): The example of the parameter.
-        kind (str): The kind of the parameter. For python function it is argument or return.
-          For OpenAPI it is location like query, path, header, cookie, body.
-    """
-
-    name: str
-    description: str = None
-    default_value: str = None
-    type_: str = None
-    type_object: Optional[Any] = Field(default=None, metadata=config(exclude=_exclude))
-    type_schema: Dict = Field(default=None)
-    required: bool = True
-    example: str = None
-    kind: str = None
-
-    def __init__(
-        self,
-        name: str,
-        description: str,
-        default_value: str = None,
-        type_: str = "str",
-        type_object: Any = None,
-        type_schema: Dict[str, Any] = None,
-        required: bool = True,
-        example: str = None,
-        kind: str = ParameterViewKind.ARGUMENT,
-    ):
-        """Initialize a ParameterView."""
-        validate_function_param_name(name)
-        self.name = name
-        self.description = description
-        self.default_value = default_value
-        self.type_ = type_
-        self.type_object = type_object
-        self.type_schema = type_schema
-        self.required = required
-        self.example = example
-        self.kind = kind
-        super().__init__()
-
-    def as_semantic_kernel_parameter(self) -> dict:
-        return {
-            "name": self.name,
-            "description": self.description,
-            "default_value": self.default_value,
-            "type": self.type_,
-            "required": self.required,
-        }
-
-    @classmethod
-    def merge(cls, a: "ParameterView", b: "ParameterView") -> "ParameterView":
-        """Merge two parameter views."""
-        return cls(
-            name=a.name,
-            description=(b.description or a.description),
-            default_value=(b.default_value or a.default_value),
-            type_=(b.type_ or a.type_),
-            required=(b.required if b.required is not None else a.required),
-            example=(b.example or a.example),
-            kind=(b.kind or a.kind),
-        )
-
-
-@dataclass
-class ManifestView(BaseModel):
-    """View of a function."""
-
-    name: str
-    description: str
-    parameters: List[ParameterView] = Field(default_factory=list)
-    returns: Optional[List[ParameterView]] = Field(default_factory=list)
-    is_async: bool = False
-    is_stream: bool = False
-    return_direct: bool = False
-
-    _name_to_params: Dict[str, ParameterView] = Field(default_factory=dict)
-
-    def __init__(
-        self,
-        name: str,
-        description: str,
-        parameters: List[ParameterView],
-        returns: Optional[List[ParameterView]] = None,
-        is_async: bool = False,
-        is_stream: bool = False,
-        return_direct: bool = False,
-    ) -> None:
-        """Initialize a ManifestView."""
-        validate_function_name(name)
-        self.name = name
-        self.description = description
-        self.parameters = parameters
-        self.returns = returns
-        self.is_async = is_async
-        self.is_stream = is_stream
-        self.return_direct = return_direct
-
-        _name_to_params = {}
-        for item in parameters:
-            _name_to_params[item.name] = item
-        self._name_to_params = _name_to_params
-
-        super().__init__()
 
 def manifest(
     *,
