+++ python/tests/test_manifest_signature.py
